+++ packages/vendure-plugin-campaign-tracker/src/api/campaign-tracker-shop.resolver.ts
@@ -3,7 +3,10 @@ import {
   Allow,
   Ctx,
   Order,
+  OrderService,
   Permission,
+  RelationPaths,
+  Relations,
   RequestContext,
   Transaction,
 } from '@vendure/core';
@@ -12,18 +15,26 @@ import { MutationAddCampaignToOrderArgs } from '../ui/generated/graphql';
 
 @Resolver()
 export class CampaignTrackerShopResolver {
-  constructor(private campaignTrackerService: CampaignTrackerService) {}
+  constructor(
+    private campaignTrackerService: CampaignTrackerService,
+    private orderService: OrderService
+  ) {}
 
   @Mutation()
   @Transaction()
   @Allow(Permission.UpdateOrder, Permission.Owner)
   async addCampaignToOrder(
     @Ctx() ctx: RequestContext,
-    @Args() { campaignCode }: MutationAddCampaignToOrderArgs
+    @Args() { campaignCode }: MutationAddCampaignToOrderArgs,
+    @Relations({ entity: Order, omit: ['aggregateOrder', 'sellerOrders'] })
+    relations: RelationPaths<Order>
   ): Promise<Order | undefined> {
-    return await this.campaignTrackerService.addCampaignToOrder(
+    const order = await this.campaignTrackerService.addCampaignToOrder(
       ctx,
       campaignCode
     );
+    if (order) {
+      return this.orderService.findOne(ctx, order.id, relations);
+    }
   }
 }